Restructure cfn decoding into a separate module
