+++ test/integration/test_good_templates.py
@@ -16,7 +16,7 @@
 """
 from cfnlint import Template, RulesCollection, Runner  # pylint: disable=E0401
 from cfnlint.core import DEFAULT_RULESDIR  # pylint: disable=E0401
-import cfnlint.cfn_yaml  # pylint: disable=E0401
+import cfnlint.decode.cfn_yaml  # pylint: disable=E0401
 from testlib.testcase import BaseTestCase
 
 
@@ -70,7 +70,7 @@ class TestQuickStartTemplates(BaseTestCase):
         for _, values in self.filenames.items():
             filename = values.get('filename')
             failures = values.get('failures')
-            template = cfnlint.cfn_yaml.load(filename)
+            template = cfnlint.decode.cfn_yaml.load(filename)
 
             runner = Runner(self.rules, filename, template, ['us-east-1'])
             matches = list()
