+++ test/module/test_rules_collections.py
@@ -16,7 +16,7 @@
 """
 from cfnlint import RulesCollection, Template
 from cfnlint.core import DEFAULT_RULESDIR  # pylint: disable=E0401
-import cfnlint.cfn_yaml  # pylint: disable=E0401
+import cfnlint.decode.cfn_yaml  # pylint: disable=E0401
 from testlib.testcase import BaseTestCase
 
 
@@ -40,7 +40,7 @@ class TestTemplate(BaseTestCase):
     def test_success_run(self):
         """ Test Run Logic"""
         filename = 'fixtures/templates/good/generic.yaml'
-        template = cfnlint.cfn_yaml.load(filename)
+        template = cfnlint.decode.cfn_yaml.load(filename)
         cfn = Template(filename, template, ['us-east-1'])
 
         matches = list()
@@ -50,7 +50,7 @@ class TestTemplate(BaseTestCase):
     def test_fail_run(self):
         """Test failure run"""
         filename = 'fixtures/templates/bad/generic.yaml'
-        template = cfnlint.cfn_yaml.load(filename)
+        template = cfnlint.decode.cfn_yaml.load(filename)
         cfn = Template(filename, template, ['us-east-1'])
 
         matches = list()
@@ -60,7 +60,7 @@ class TestTemplate(BaseTestCase):
     def test_fail_sub_properties_run(self):
         """Test failure run"""
         filename = 'fixtures/templates/bad/properties_onlyone.yaml'
-        template = cfnlint.cfn_yaml.load(filename)
+        template = cfnlint.decode.cfn_yaml.load(filename)
         cfn = Template(filename, template, ['us-east-1'])
 
         matches = list()
